Move DSACryptoServiceProvider tests to S.S.C tests and share ShimHelpers
- Moved DSACryptoServiceProviderTests.cs from S.S.C.Csp/tests to S.S.C/tests
  with file lists in both the Windows and Unix ItemGroups so it runs on
  the two platforms where DSACryptoServiceProvider has a real implementation.
- Renamed DSACryptoServiceProviderProvider.Windows.cs and
  DSATestRegistration.Csp.Windows.cs to drop the .Windows suffix since they
  are no longer Windows-only.
- Split ShimHelpers into a partial class: TestSymmetricAlgorithmProperties
  stays in the Csp tests project; VerifyAllBaseMembersOverridden moves to
  S.S.C/tests/ShimHelpers.cs and is linked into the Csp tests project so
  it continues to be reachable from the symmetric algorithm tests.
